issue: 327792 GRO delayed ack issue - netperf client finishes with "Connection reset by peer"
log:
with GRO enabled:
the sender use all its intial congestion window to send data (the test send large packets).
the receiver batch all this data to one packet and bring it to the tcp stack.
the tcp stack will not ack this packet due to ack delay mechanism, waiting for the next packet to arrive.
the next packet does not arrive because the sender congestion window is full.
after 10 msec, the timer of the receiver force sending the ack.
this will continue to happen until the congestion window of the sender will be large enough
for sending amount of data which is larger than the amount that the receiver can batch to one packet.
this can substaintaly slow down the the TCP slow start phase, and get everything stuck.

solution: do not delay ack if it is batched packet.

reviewed by: <Name>

Change-Id: I9f6e77b518592039fe3db15c42865ea7bfd57bf7
Signed-off-by: ork <ork@mellanox.com>