issue: 351353 wakeup mechanism is not fit for multithread access to the same fd
log:
should be used as a semaphore not as a mutex.
remove the wake-up pipe after all waiting threads on the fd have been
woken-up.

Also, sched_yield in do_wakeup() currently does not benefit at all, and
actually make things worse.
it is always called from inside a lock which is needed by the waking
thread, therefore the scheduler forced to do context-switches back and
forth between threads for no reason.

reviewed by: AlexR

Change-Id: I41d18e7d51e872d9b9b9b2c42d1689e2b0c179c6
Signed-off-by: ork <ork@mellanox.com>