issue: 355605 create timers groups to lower to load of handling many timers
log:
the group is registered to the internal thread timer, while the handler
timers are registered to the group.
the group can have its own logic of triggering the timers and to have
efficient add/remove timer handlers.

we need it for TCP sockets.
when we have thoushands of sockets, each registering a timer, it can be more
efficient handling them in a timer group.

reviewed by: AlexR

Change-Id: I7c7977b7c7f13d082253a120c564c81cc5c10af0
Signed-off-by: ork <ork@mellanox.com>