issue: 317 Loopback support on Ethernet (Multicast & Unicast)
log:
As OFED 2.1-0.0.5 is now supporting Eth loopback, we can remove the
limitation of not-offloading ETH loopback.
ETH loopback will be offloaded and will work only between VMA processes,
or when the receiver is offloaded.
If it need to work with non-offloaded receiver, need to use libvma.conf
rule to not offload this loopback.

Change-Id: Ib97009124c624abc67730013dfeaa7e9195acfa5
Signed-off-by: ork <ork@mellanox.com>