issue: 375244 Align VMA with latest MLNX_OFED 2.2 experimental verbs
log:
MLNX_OFED 2.2 is now in sync with the linux kernel upstream code.
All verbs capabilites that are not yet in the kernel upstream moved to
verbs_exp.h as experimental verbs.
Changed VMA code that use these capabilites to use the experimental
verbs while maintaining the ability to compile with older MLNX_OFED
versions.

reviewed by: AlexR

Change-Id: I3feee4d041ed5a8a1b9e64e55220a9d9dc8ffe28
Signed-off-by: Or Kehati <ork@mellanox.com>