issue: 434108 extract lwip - apply patch 022-fix_connect_with_SO_REUSEADDR_option.patch
log:
This patch fixes a bug in connect() when SO_REUSEADDR is enabled.
When calling to bind() before connect() the pcb is added the bounded
list.
when SO_REUSEADDR is enabled we confirm that the 5 tuple is unique.
because the pcb is in the bounded list we will always find a match (same
pcb) and return ERR_USE.

Change-Id: I64903579212755a5dabd3b2a3f35dcfa7624af7e
Signed-off-by: Or Kehati <ork@mellanox.com>