issue: 434108 extract lwip - apply patch 030-update_rexmit_handling.patch
log:
VMA needs to know if lwip sends a packet for retransmission,
because VMA give back the buffer to lwip before it was actually sent.
Then, in case that an ack for this packet have arrived to lwip
before we actually sent the retransimisson, lwip will reuse the
buffer and we will send garbage by accident.
In order to avoid such cases, ip_output have been changed to let
VMA know that it retransmit a packet, so VMA can make a copy of
the buffer before returning to lwip.

Change-Id: Ib27e24f75fd0604215449877ee5251579ccd37c7
Signed-off-by: Or Kehati <ork@mellanox.com>