issue: 618523 LAG bond receive double amount of multicast traffic if LACP not working
This is due to the switch sending on both ports, and us putting flow
steering rules on both.
The OS is polling only one port if the LACP not working (= port
state is not active).
So we should also use only active ports according to the file:
/sys/class/net/<slave_if_name>/bonding_slave/state
If no port is LACP active, but some are up, we make the first one
active.
The fix is only for kernel > 3.14 or RH7.2 and higher (due to
present of the state file).

Change-Id: I4a6ee663aa029a23943d609a9438489202db9da4
Signed-off-by: Or Kehati <ork@mellanox.com>