issue: 647962 Fix a crash when all bond's slaves ports are taken down and then up
The crash happens due to TX buffers which are freed twice.
The reason is a pointer to a previously sent packets (sent during
port is down) which is not initialize when port is up, and linked
to new packets that are sent when port up.
Also, if port is down, avoid sending on this port, so we won't get
to this scenarios in the first place, and save cpu cycles.

Change-Id: I0f401bb0f78cc2b937f32200dac8cfa01acd467c
Signed-off-by: Or Kehati <ork@mellanox.com>