issue: 781540 wait for Rx WQE's when closing RING (QP)
The handling of all FLUSHed WQE's to CQE's takes a different amount of time on each HW.
This commit mark's the last posted WQE. Later, when QP is moved to ERROR state, the
qp_mgr::release_rx_buffers() knows to wait for that marked WR_ID to indicate all CQE's are FLUSHed.
Tests:
1. server_socket_recreate_loop.py
simple open and close of UDP socket which creates and releases a RING in a loop
are returned to poll as expected.
2. server_socket_receive_and_recreate_loop.py
open and close multiple UDP sockets which creates and releases a RING in a loop
and expects to receive some traffic on any of the opened sockets
Using 'vma_stats -v 3' you can follow the RX buffer pool and verify all Rx buffers

Signed-off-by: Alex Rosenbaum <rosenbaumalex@users.noreply.github.com>
Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>