issue: 844374 Fix error handling while register socket with plural epoll instances
VMA does not supports registration of socket with plural epoll instances.
In the current situation we simply ignore the request without returning
an error.
errno will be set to EEXIST if the fd is already registered with the same epoll.
errno will be set to ENOMEM if the fd is already registered with another epoll.