issue: 884440 Fix handling of SO_LINGER in case l_linger is zero
If l_onoff is nonzero and l_linger is zero, TCP aborts the connection
when it is closed.

According to "UNIX Network Programming" third edition,
setting SO_LINGER with timeout 0 prior to calling close()
will cause the normal termination sequence not to be initiated.
If l_onoff is nonzero and l_linger is zero, TCP aborts the connection when it is closed.
That is, TCP discards any data still remaining in the socket
send buffer and sends an RST to the peer, not the normal four-packet connection
termination sequence

Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>