issue: 768398 Add daemon
This service is implemented to handle a scenario known as
a “half-open connection”.
This scenario is called a “half-open connection” because one side
realizes the connection was lost but the other side believes it is
still active.

Special service is started during OS initialization or manually
under super user permissions. The purpose of one is collecting
information about TCP connections from all VMA launches, detect
termination of them and notification their peers about process
destruction.

Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>