issue: 1011829 Add VMA_TCP_NODELAY parameter (Nagle algorithm)
By default Nagle algorithm is enabled for newly created TCP sockets.
If this parameter is enabled Nagle algorithm will be disabled for newly
created TCP sockets.
This means that TCP segments are always sent as soon as possible, even if
there is only a small amount of data.
For more information on TCP_NODELAY flag refer to TCP manual page.
This parameter is enabled when setting VMA_SPEC to latency or
ultra-latency.
Default value is disabled.

Signed-off-by: Liran Oz <lirano@mellanox.com>