issue: 1037215 rdma_lib_reset not detected
On some gcc versions e.g. 5.4.0 (Ubuntu 16.04)
there is an issue with the order of arguments.
The -l must be used after the name of the source
file or the function won't be detected.
Replacing AC_CHECK_LIB with AC_SEARCH_LIBS adds -l
after the source file and by that autotools detects
rdma_lib_reset.

Signed-off-by: Rafi Wiener <rafiw@mellanox.com>