issue: 1037215 Fix autotool detection of rdma_lib_reset
In some gcc versions e.g. 5.4.0 (Ubuntu 16.04) there is an issue with
the order of arguments.
When linking with a library function, the -l<library> must be placed
after the source file to which it is being linked otherwise the
function won't be detected when linking.

Replacing AC_CHECK_LIB with AC_SEARCH_LIBS places -lrdmacm after
autotool's test source file conftest.c - allowing rdma_lib_reset to be
found and linked (see the config.log resulting from running ./configure).

Signed-off-by: Rafi Wiener <rafiw@mellanox.com>