issue: 1062367 Fix TCP window deadlock in case of a small window
The deadlock occurs when the head of TCP unsent queue packet length
is bigger than the send window combined with loss of window update
message that sent from peer side (ACK segment that contain no data).

ACK segment that contain no data are not reliably transmitted by TCP.
Therefore, VMA may hang when an ACK segment that update the window
is lost.

The solution is to split and fill the remaining window with part of
the unsent segment (which will engage zero-window probing upon
reception of the zero window update from the receiver).

For more details, see RFC-1122.

Signed-off-by: Daniel Libenson <danielli@mellanox.com>