issue: 1156840 Add TCP Tx window availability
While enabled, TCP packets will only be sent if their size
(hdr options + data) is less than or equal to the window size.
Otherwise -1 is returned and errno is set to EAGAIN.

In case of using writev() with iov, only the first buffer will
be tested (iov[0]) for window availability criteria.

To enable:
./configure --enable-tcp-tx-wnd-availability

Signed-off-by: Daniel Libenson <danielli@mellanox.com>