issue: 1164732 Improve thread safity of neigh_table_mgr
There is no need to protect neigh_table_mgr object in this place.
I assume that initial motivation to use lock in this place was a risk
of conflict between create_new_entry() and entry search.
So just move
g_p_netlink_handler->register_event(nlgrpNEIGH, this)
in constructor as for
g_p_netlink_handler->register_event(nlgrpROUTE, this)

Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>