issue: 1179641 Add parsing route table metrics
This commit adds parsing of rotueing metrics to vma.
Currently only RTAX_MTU and RTAX_LOCK are supported.
VMA uses the mtu to calculate the mss.
VMA only supprots lock mode due to the face mtu updates
are done using ICMP packet that VMA is not capturing.
when using RTAX_LOCK VMA will use mtu of 536 as defined
in RFC 879.

Signed-off-by: Rafi Wiener <rafiw@mellanox.com>