issue: 1176937 Add TCP_QUICKACK option support
Enable set/get quick ack mode possibility.
Enable quickack mode if set or disable quickack mode if cleared.
In quickack mode, acks are sent immediately, rather than delayed
if needed in accordance to normal TCP operation. This flag is
not permanent, it only enables a switch to or from quickack mode.
In VMA internal implementation this mode is sticky. Once it is
set as on/off the mode can be applied until next explicit
direction.

Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>