issue: 1185372 Fix traffic is not offloaded for high VMA_RX_WRE values
As part of VMA's initialization process, we are trying to open QP on
each ETH device (verify_eth_qp_creation()).
This fails when VMA is trying to open QP using high value of
VMA_RX_WRE which is not supported.
A better way to handle this is to use a default value of VMA_RX_WRE at
verify_eth_creation().

Signed-off-by: Liran Oz <lirano@mellanox.com>