issue: 1248968 Add sesockopt for IP_TOS
This commit adds support for IP_TOS attribute in setsockopt,
witch is written inside the IP TOS field.

VMA had a lack of usage in the tos when resolving the route
and used the route mtu in places it shouldn't had.
This commit adds the TOS and PCP and alignes all the tos
usage

Signed-off-by: Rafi Wiener <rafiw@mellanox.com>