issue: 1226478 Add direct-ring creation and HW descriptors logic
This commit adds a new ring type called direct ring.
This ring allows an application direct access to the HW descriptors
that the ring uses (e.g. send queue an completion queue).
This allows the application to send and receive date without VMA
managing the memory.
With this ring you cannot use any socket API on the socket fd.

Signed-off-by: Rafi Wiener <rafiw@mellanox.com>