issue: 1262144 Print warning when extra api functions are not supported
VMA extra api struct is not initialized with unsupported functions.
For example, when socketXtreme was not enable during configuration time
and when Striding RQ is not supported,
This can cause the user application to crash.
This commit will avoid those crashes by set the errno to EOPNOTSUPP and
print a relevent warning message.

Signed-off-by: Liran Oz <lirano@mellanox.com>