issue: 1426871 Add debian watch file
This is an optional, recommended configuration file for the uscan
utility which defines how to automatically scan ftp or http sites
for newly available updates of the package. This is also used by
some Debian QA tools to help with quality control and maintenance
of the distribution as a whole.

Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>