issue: 1734068 Fix inaccurate timer_handler unregistration
Special protection is needed to avoid scenario when deregistration is done
during timer_handler object destruction, timer node itself is not removed
and time for this timer node is expired. In this case there is no guarantee
to operate with timer_handler object.
So assign orig_time_msec field to maximum value;

Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>