issue: 1606044 Optimize sendfile() for TCP
This sendfile optimization is based on doing data copying
directly to pbuf on lwip layer.
Single call of readv() is done for several pbufs.
Effective operation should do read from storage with page sized
blocks. So it is recomended to use VMA_TX_BUF_SIZE as 4096 or 8192.

The same can be done using mmap approach.
Memory mapping a file directly avoids copying buffers which happen
with read() and write() calls. Calls to read() and write() include
a pointer to buffer in process' address space where the data is stored.
Kernel has to copy the data to/from those locations. Using mmap() maps
the file to process' address space, so the process can address the
file directly and no copies are required.
There is also no system call overhead when accessing memory mapped file
after the initial call.
But there are (-):
- mmap should be aligned by page size
- mmaped buffer should be protected from file truncate scenario

Issues:
2044752 Add check for count parameter of sendfile()
2044574 Fix current buffer offset position
2056110 Fix failure return errno

Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>