issue: 2136792 Remove IS_RELEASE_ZERO usage
IS_RELEASE_ZERO is used to provide soname for binary with
release number.
It is needed to include release number in library name
for example libvma-1.so.

By specifying the release number with this option,
libtool will build a library that does this, but will break
binary compatibility for each change of the release number.
By breaking binary compatibility this way, you negate the
possibility of fixing bugs in installed programs by installing
an updated shared library.
You should probably be using ‘-version-info’ instead.

so let do it using -version-info option with name format as
libvma.so.9.0.2 for example

Signed-off-by: Igor Ivanov <igor.ivanov.va@gmail.com>