issue: 2175248 Check snprintf() return value
Handle negative return code to suppress warnings from static analyzers.
Cache some constant string in m_str instead of ignoring the error in
order not to call snprintf() on each iteration.

We don't expect negative error code since we don't work with wide
characters encoding here. Also, VMA doesn't work on systems with glibc
that returns -1 for truncated strings.

We don't have to handle truncated strings, because this is not an issue.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>