issue: 2394155 Remove shutdown slave ring from xmit_rings
In RoCE LAG mode we keep the first slave ring active for RX in order
to avoid duplicate steering rules and packet duplications as result.
Field m_active remained true because of this change and it is used for
building the array of active transmit rings.

Since we cannot send through a down ring, we have to remove such a ring
from the m_xmit_rings array. Set m_active field to the netdev active
status and use it to check whether we can use it for TX.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>