issue: 2740920 Add support for fortified glibc functions
The FORTIFY_SOURCE macro provides lightweight support for detecting buffer overflows
in various functions that perform operations on memory and strings. Not all types of
buffer overflows can be detected with this macro, but it does provide an extra level
of validation for some functions that are potentially a source of buffer overflow flaws.
It protects both C and C++ code. FORTIFY_SOURCE works by computing the number of bytes
that are going to be copied from a source to the destination.

If _FORTIFY_SOURCE is set to 1, with compiler optimization level 1 (gcc -O1) and above,
checks that shouldn't change the behavior of conforming programs are performed.
With _FORTIFY_SOURCE set to 2  some  more  checking  is added, but some conforming
programs might fail.  Some of the checks can be performed at compile time, and result
in compiler warnings; other checks take place at run time, and result in a run-time
error if the check fails.  Use of this macro requires compiler support,
available with gcc(1) since version 4.0.

__read_chk (GLIBC_2.4)
__recv_chk (GLIBC_2.4)
__recvfrom_chk (GLIBC_2.4)
__poll_chk (GLIBC_2.16)
__ppoll_chk (GLIBC_2.16)

Signed-off-by: Igor Ivanov <igori@nvidia.com>