issue: 3092555 Cont for PR992: Blocking socket connect race
In case of blocking connect failure, the timer for the socket was already been added.
However, when blocking socket fails the dest-entry is destroyed.
Since the timer was alrady registered it may try send a SYN rexmit,
but will fail with seg-fault since the dest entry had been destroyed.
As a fix, we set tcp-state to close.
As part of the fix we also do clean handling of second connect() attempt on the same socket,
which already failed a connect. In this case we return ECONNABORTED as this scenario is
currently unsupported.
From conect() manual https://man7.org/linux/man-pages/man2/connect.2.html:
"If connect() fails, consider the state of the socket as unspecified. Portable applications
should close the socket and create a new one for reconnecting."

Signed-off-by: Alexander Grissik <agrissik@nvidia.com>