issue: 4337575 Init VMA in vma_thread_offload()
vma_get_api()->thread_offload() can be called before the 1st socket()
syscall. In this case VMA is in uninitialized state and the
thread_offload() call fails with -1.

Call DO_GLOBAL_CTORS() in the vma_thread_offload() to guarantee the
initialized state in all scenarios.

Until VMA 9.8.70, the initialization was done in the vma_get_api() call,
however, it was removed due to external allocator support which requires
uninitialized state to guarantee buffers allocation with the configured
allocator.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>