[CI] issue: 4377705 Create tar.gz from src.rpm
Released packed should be tar.gz instead of src.rpm, with a small change
to the dir stucture

Move libvma.spec to the root of the directory being compressed.
Remove libvma.spec from contrib/scripts/libvma.spec

For now we will create both the src.rpm and the tar.gz
created in the build_pkg.sh , and moves it to the release folder.

Pointers to the new tar.gz file are not being created by the script

Signed-off-by: NirWolfer <nwolfer@nvidia.com>